Add QR Scanner dialog
