+++ packages/yoroi-extension/app/components/wallet/send/WalletSendFormSteps/QRScannerDialog.scss
@@ -1,0 +1,18 @@
+.component {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    width: 100%;
+    height: 100%;
+    .scanner {
+        display: inline-block;
+        width: 300px;
+        height: 300px;
+    }
+}
+
+
+.dialog {
+    min-width: 380px !important;
+    min-height: 425px !important;
+}
