+++ packages/yoroi-extension/app/i18n/locales/en-US.json
@@ -166,6 +166,7 @@
   "wallet.send.form.dialog.noTokensYet": "There are no tokens in your wallet yet",
   "wallet.send.form.dialog.nNft": "NFT ({number})",
   "wallet.send.form.dialog.nToken": "Tokens ({number})",
+  "wallet.send.form.qrDialog.title": "Scan QR code",
   "global.labels.processing": "Processing...",
   "global.labels.remove": "Remove",
   "global.labels.rewardsLabel": "Rewards",
