+++ packages/yoroi-extension/package.json
@@ -220,7 +220,6 @@
     "react-dom": "17.0.2",
     "react-intl": "2.9.0",
     "react-markdown": "5.0.3",
-    "react-qr-reader": "^3.0.0-beta-1",
     "react-router": "5.2.0",
     "react-router-dom": "5.2.0",
     "react-webcam-qr-scanner": "^1.0.8",
